requirements.txt
Django==3.1
psycopg2-binary==2.8.5
Pillow==7.1
python-decouple
django-crispy-forms
django-shopping-cart
decouple


{% load static %}
<!DOCTYPE html>
<html>
  <head>
    <meta charset="utf-8" />
    <title>{% block title %}CaseWorld{% endblock %}</title>
    <link href="{% static "css/base.css" %}" rel="stylesheet">
    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.1/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-F3w7mX95PdgyTmZZMECAngseQB83DfGTowi0iMjiWaeVhAn4FJkqJByhZMI3AhiU" crossorigin="anonymous">
  </head>
  <body>

  {% if not request.user.is_authenticated %}

            <a href="{% url 'register' %}">Sign Up</a>
            <a href="{% url 'login' %}">Log in</a>
            {% else %}
            <h4>{{ request.user.username}}</h4>
            <a href="{% url 'logout' %}">Log out</a>
          {% endif %}

  <div id="header">
      <a href="/" class="logo">CaseWorld</a>
    </div>
    <div id="subheader">
      <div class="cart">
        {% with total_items=cart|length %}
          {% if total_items > 0 %}
            Your cart:
            <a href="{% url "cart:cart_detail" %}">
              {{ total_items }} item{{ total_items|pluralize }},
              ${{ cart.get_total_price }}
            </a>
          {% else %}
            Your cart is empty.
          {% endif %}
        {% endwith %}
      </div>
    </div>
    <div id="content">
      {% block content %}
      {% endblock %}
    </div>
  </body>
</html>
